Crate radio

This crate utilizes SDRs for long range communications intended for space bound communications. This crate will allow for building more code on top of and make it easier for networking based protocols to be developed and implemented with use of an SDR.

It should be noted that any library that uses this crate will need to run with root privileges or else the system won’t be able to connect to the radio.

Modules

dsp

Digital Signal Processing (DSP) is the hallmark and a widely studied field in radio communications and is

frame

This file contains the Frame object struct and implementation. Frame is the lowest OSI level building block which is best handed and managed by the system itself. This particular frame header implements CCSDS SDLP standard which is the standard frame header for space communication

rx_handling

This file holds RX operations and it’s various helper functions. This improves the QoL and readability of the code which is essential for understanding the complex operations that take place during the read thread’s loop.

tools

This file holds various functions and helper methods that improve the over all QOL of working in this crate

Structs

RadioStream

Radio stream is the main class of the Radio crate. This makes it easy for anyone to RX or TX with an SDR from Rust. This is done by providing RX and TX methods through this object

Enums

ModulationType

This is a list of different signal modulation types currently accepted by the radio

Statics

AMBLE

This set of bits is sent ahead of the real transmission frame to allow time for the SDR to sync and better be able to read the frame that follows.

IDENT

IDENT (AKA identifier) is a set of bits that allows the radio to know that it has reached the end of the preamble and is about to read the frame data.

MOD_TYPE

This is the current Modulation scheme that the radio is configured to accept

Functions

bits_per_symbol

This will return the number of bits being currently transmitted per a symbol

demodulation

This is a helper function that makes it easy to change mod and demod functions on the fly.

modulation

This is a helper function that makes it easy to change mod and demod functions on the fly.