Struct radio::RadioStream

pub struct RadioStream {
    pub tx_stream: Tx,
    pub modulation: Modulators,
    pub rx_buffer: Arc<RwLock<Vec<Vec<u8>>>>,
    pub settings: RadioSettings,
}

Radio stream is the main class of the Radio crate. This makes it easy for anyone to RX or TX with an SDR from Rust. This is done by providing RX and TX methods through this object

Fields
tx_stream: Tx

The tx stream that data is passed to to have transmitted out of the radio

modulation: Modulators

This is a pre-built modulation object that allows for quick modulation of out going data

rx_buffer: Arc<RwLock<Vec<Vec<u8>>>>

This is a thread safe buffer that the RX thread will fill with received transmissions that can be read from the read method

settings: RadioSettings

This is a saved copy of the radio settings that can be altered later depending the needs and wants of the user.

Implementations

impl RadioStream

pub fn new() -> Result<RadioStream>

This will create a new radio stream object.

This will return an error if no radio is detected or can be connected to.

pub fn transmit(&self, data: &[u8]) -> Result<()>

This will transmit binary data to the radio

pub fn transmit_frame(&self, frame: &Frame) -> Result<()>

This a wrapper function that allows for direct transmission of frames

pub fn read(&self) -> Result<Vec<u8>>

This process samples read and return any data received

pub fn receive_frames(&self) -> Result<Vec<Frame>>

This is a wrapper class that allows for the reading and receiving of frames directly